add crud for posts. refactor blogs crud

+++ src/modules/bloggers-platform/blogs/api/blogs.controller.ts
@@ -4,6 +4,7 @@ import {
   Delete,
   Get,
   HttpCode,
+  HttpStatus,
   NotFoundException,
   Param,
   Post,
@@ -13,12 +14,10 @@ import {
 import { BlogsService } from '../application/blogs.service';
 import { GetBlogsQueryParams } from './input-dto/get-blogs-query-params';
 import { BlogsQueryRepository } from '../infrastructure/blogs-query.repository';
-import { IBasePaginatedView } from 'src/core/types/base-paginated-entity-view';
 import { BlogsViewDto } from './view-dto/blogs-view-dto';
 import { CreateBlogDto } from './input-dto/create-blog.dto';
-import { MongooseObjtId } from 'src/core/types/mongoose-objectId';
 import { UpdateBlogDto } from './input-dto/update-blog.dto';
-import { HTTP_STATUS_CODES } from 'src/core/http-status-codes';
+import { BasePaginatedView } from 'src/core/dto/base-paginated-view';
 
 @Controller('blogs')
 export class BlogsController {
@@ -30,14 +29,12 @@ export class BlogsController {
   @Get()
   async getAllBlogs(
     @Query() query: GetBlogsQueryParams,
-  ): Promise<IBasePaginatedView<BlogsViewDto>> {
+  ): Promise<BasePaginatedView<BlogsViewDto>> {
     return await this.blogsQueryRepository.getAllBlogs(query);
   }
 
   @Get(':blogId')
-  async getBlogById(
-    @Param('blogId') blogId: MongooseObjtId,
-  ): Promise<BlogsViewDto> {
+  async getBlogById(@Param('blogId') blogId: string): Promise<BlogsViewDto> {
     const blog = await this.blogsQueryRepository.getBlogById(blogId);
 
     if (!blog) {
@@ -48,23 +45,43 @@ export class BlogsController {
   }
 
   @Post()
-  async createBlog(@Body() blog: CreateBlogDto): Promise<BlogsViewDto | null> {
+  async createBlog(@Body() blog: CreateBlogDto): Promise<BlogsViewDto> {
     const blogId = await this.blogsService.createBlog(blog);
-    return await this.blogsQueryRepository.getBlogById(blogId);
+    const newBlog = await this.blogsQueryRepository.getBlogById(
+      blogId.toString(),
+    );
+
+    if (!newBlog) {
+      throw new NotFoundException('blog not found');
+    }
+
+    return newBlog;
   }
 
   @Put(':blogId')
-  @HttpCode(HTTP_STATUS_CODES.NO_CONTENT_204)
+  @HttpCode(HttpStatus.NO_CONTENT)
   async updateBlog(
-    @Param('blogId') blogId: MongooseObjtId,
-    @Body() blog: UpdateBlogDto,
+    @Param('blogId') blogId: string,
+    @Body() updates: UpdateBlogDto,
   ): Promise<void> {
-    await this.blogsService.updateBlog(blogId, blog);
+    const blog = await this.blogsQueryRepository.getBlogById(blogId);
+
+    if (!blog) {
+      throw new NotFoundException('blog not found');
+    }
+
+    await this.blogsService.updateBlog(blogId, updates);
   }
 
   @Delete(':blogId')
-  @HttpCode(HTTP_STATUS_CODES.NO_CONTENT_204)
-  async deleteBlog(@Param('blogId') blogId: MongooseObjtId): Promise<void> {
+  @HttpCode(HttpStatus.NO_CONTENT)
+  async deleteBlog(@Param('blogId') blogId: string): Promise<void> {
+    const blog = await this.blogsQueryRepository.getBlogById(blogId);
+
+    if (!blog) {
+      throw new NotFoundException('blog not found');
+    }
+
     await this.blogsService.deleteBlog(blogId);
   }
 }
